backend/requirements.txt
flask
firebase_admin
werkzeug
Flask-JWT-Extended
python-dotenv
functions-framework==3.*